Split up lobby components
