+++ src/routes/api/createLobby/+server.ts
@@ -19,7 +19,8 @@ export const POST = (async ({ request }) => {
     // add new user to firestore
     await setDoc(doc(db, 'lobby', uid), {
       gameMode: 'solo',
-      listOfUsers: [uid]
+      listOfUsers: [uid],
+      status: 'SETUP'
     });
 
     return json(true);
